requirements.txt
akshare>=1.14.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
feedparser>=6.0.0
readability-lxml>=0.8
markdownify>=0.12
